Combine InitializationOnDbItSpec
